+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Increase Custom Labels value max length to 500. (UIEH-985)
 * Fix routing issues when visiting Note Edit page. (UIEH-991)
 * Add permission for button New Settings (eHoldings): Can create, edit, and view knowledge base credentials. (UIEH-992)
+* Title Record: Add a Selection Status filter to Search within Packages modal. (UIEH-978)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)